Create Scala
In this search, we use a random number generator Randomly to guess numbers from 1 to 100. The player must guess this number by entering the received number using the console. The program suggests a calculated relationship, is the estimated too small or too large, and counts the number of consumed quantities used to guess the volumes of the number. After guessing the number of players, play again.